Attention: cast scale to np.float32 to keep chain in F32
* The previous code computed `scale = 1.0 / np.sqrt(head_dim)`
  which is a Python float (i.e. float64). When the F32
  attention tensor is multiplied by a Python float,
  NEP 50 promotes the entire result to float64.

* This leaked F64 throughout the rest of the attention
  path: `qk_matmul` output, `causal_mask` add, softmax
  intermediates, `av_matmul` output, the post-attention
  reshape, and CRITICALLY the residual `hidden = residual
  + attn_out` which became F64 — meaning every MLP weight
  matmul downstream ran on F64 hidden state and produced
  F64 output that propagated further.

* Verified in the per-projection audit: `attn_pre_mask`,
  `attn_masked`, `attn_probs`, and `context` were all
  dtype=float64 shape=(16, 1, 121) nbytes=15.1KB. After
  the fix they are dtype=float32 nbytes=7.6KB — half the
  bandwidth for every attention intermediate and every
  hidden-state vector that crosses a residual boundary.

* Cast `scale` once at the top of each decoder block
  (qwen2/qwen3 `_decoder_block`, gemma3
  `_gemma3_decoder_block`, qwen3.5
  `_qwen35_full_attention_block`). Also caught the
  equivalent site inside `forward()` for the post-attention
  final norm + LM head path. Three call sites total —
  one per arch-block + one in `forward()`.

* 168 tests still pass.

+++ src/flatrun/model/qwen2.py
@@ -1415,7 +1415,7 @@ def make_qwen2_forwarder(
             k_full = np.repeat(k_hist, head_group, axis=1)
             v_full = np.repeat(v_hist, head_group, axis=1)
 
-        scale = 1.0 / np.sqrt(head_dim)
+        scale = np.float32(1.0 / np.sqrt(head_dim))
         with _p("qk_matmul", _P):
             attn = np.einsum("thd,Thd->htT", q, k_full) * scale
         with _p("causal_mask", _P):
@@ -1616,7 +1616,7 @@ def make_qwen2_forwarder(
         k_full = np.repeat(k_hist, head_group, axis=1)
         v_full = np.repeat(v_hist, head_group, axis=1)
 
-        scale = 1.0 / np.sqrt(attn_scale)
+        scale = np.float32(1.0 / np.sqrt(attn_scale))
         attn = np.einsum("thd,Thd->htT", q, k_full) * scale
         attn = attn + _causal_mask(seq_len, past_len, config.sliding_window)
         # Gemma 2/3 default soft-cap (50.0) prevents the Q.K product
@@ -1776,7 +1776,7 @@ def make_qwen2_forwarder(
         k_full = np.repeat(k_hist, head_group, axis=1)
         v_full = np.repeat(v_hist, head_group, axis=1)
 
-        scale = 1.0 / np.sqrt(head_dim)
+        scale = np.float32(1.0 / np.sqrt(head_dim))
         attn = np.einsum("thd,Thd->htT", q, k_full) * scale
         attn = attn + _causal_mask(seq_len, past_len, config.sliding_window)
         attn = _softmax(attn, axis=-1)
